l1/requirements.txt
numpy
scikit-learn

python-dotenv

langchain-huggingface
transformers
huggingface-hub

langchain-google-genai
google-generativeai

langchain-anthropic

langchain-openai
openai

langchain
langchain-core

